[SPARK-41423][CORE] Protobuf serializer for StageDataWrapper
